Add sponsors and jobs
Added crud and admin for sponsors and jobs.

This includes:

    - Placeholder templates (not styled)
    - Admin interface
    - Models
    - Migrations

Also, added django filebrowser for automatic resizing of sponsor
logos, and images in general. Once we decide on the size we can tell
filebrowser to pre-generate those sizes and then we can use them in
templates. Also filebrowser is good allowing admin users to manage media
files via the admin interface.

Added grappelli as well. Skin for the admin interface. Filebrowser
requires it.