Adds ability to have co_presenter on a talk.
Since this would be a pain to refactor properly and I don't have the
time currently I've just added a ForeignKey field to the talk model.

This let's us have just 2 people (the main speaker that submited the
paper application) and the co-speaker. If for some reason we would need
more then that we can always add a third field but I think this is
reasonable for now.